requirements.txt
black ~= 22.6
pandas ~= 1.4
mypy ~= 0.99
pylint ~= 2.15
anytree ~= 2.8
requests ~= 2.28
types-requests ~= 2.28
pyyaml ~= 6.0
types-pyyaml ~= 6.0
